+++ internal/services/macos_onboarding_settings/resource_schema.go
@@ -15,6 +15,14 @@ func ResourceJamfProMacOSOnboardingSettings() *schema.Resource {
 		ReadContext:   readWithCleanup,
 		UpdateContext: update,
 		DeleteContext: delete,
+		SchemaVersion: 1,
+		StateUpgraders: []schema.StateUpgrader{
+			{
+				Type:    resourceMacOSOnboardingSettingsV0().CoreConfigSchema().ImpliedType(),
+				Upgrade: upgradeMacOSOnboardingSettingsV0toV1,
+				Version: 0,
+			},
+		},
 		Timeouts: &schema.ResourceTimeout{
 			Create: schema.DefaultTimeout(1 * time.Minute),
 			Read:   schema.DefaultTimeout(1 * time.Minute),
@@ -28,7 +36,7 @@ func ResourceJamfProMacOSOnboardingSettings() *schema.Resource {
 				Description: "Enable or disable macOS onboarding",
 			},
 			"onboarding_items": {
-				Type:        schema.TypeList,
+				Type:        schema.TypeSet,
 				Optional:    true,
 				Description: "List of onboarding items to display during device setup",
 				Elem: &schema.Resource{